requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
plotly>=5.15.0
python-dotenv>=1.0.0
alpha-vantage>=2.3.1
requests>=2.31.0
